services/intent_gateway/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pyyaml==6.0.2
httpx==0.27.2
